refactor extract useReactivateAction
